+++ components/script/dom/mod.rs
@@ -245,8 +245,8 @@ pub(crate) mod console;
 pub(crate) mod cookiestore;
 pub(crate) mod countqueuingstrategy;
 mod create;
-pub(crate) mod credential;
-pub(crate) mod credentialscontainer;
+pub(crate) mod credentialmanagement;
+pub(crate) use self::credentialmanagement::*;
 pub(crate) mod crypto;
 pub(crate) mod cryptokey;
 pub(crate) mod csp;
@@ -382,7 +382,6 @@ pub(crate) mod notification;
 pub(crate) mod pagetransitionevent;
 pub(crate) mod paintsize;
 pub(crate) mod paintworkletglobalscope;
-pub(crate) mod passwordcredential;
 pub(crate) mod performance;
 #[allow(dead_code)]
 pub(crate) mod performanceentry;
